removed profiletab
