+++ src/components/widgets/CardEvent/CardEvent.module.css
@@ -42,7 +42,7 @@
     line-height: 1.38rem;
     text-align: left;
     display: -webkit-box;
-    -webkit-line-clamp: 2;
+    -webkit-line-clamp: 3;
     -webkit-box-orient: vertical;
     overflow: hidden;
     text-overflow: ellipsis;
