+++ src/components/widgets/EventList/EventList.module.css
@@ -2,5 +2,6 @@
     width: 100%;
     display: flex;
     flex-wrap: wrap;
+    row-gap: 4.5rem;
     column-gap: 1.25rem;
 }
