--- src/components/widgets/ProfileTabs/index.tsx
@@ -1,1 +1,0 @@
-export { ProfileTabs } from "./ProfileTabs";
